warehouse DCM now perl, not Pro*C.

+++ moira/gen/warehouse.sh
@@ -1,10 +1,5 @@
 #!/bin/sh
-#
-#
-DIR=/warehouse/transfers/moirausr
-TAR_FILE=$DIR/moira.tar.Z
-
-PATH="/etc:/bin:/usr/bin:/usr/etc:/usr/athena/etc"
+# $Id: warehouse.sh,v 1.5 2007-01-03 20:31:59 zacheiss Exp $
 
 if [ -d /var/athena ] && [ -w /var/athena ]; then
     exec >/var/athena/moira_update.log 2>&1
@@ -14,26 +9,20 @@ fi
 
 # The following exit codes are defined and MUST BE CONSISTENT with the
 # error codes the library uses:
-set MR_MISSINGFILE = 	47836473
-set MR_TARERR = 	47836476
+MR_MISSINGFILE=47836473
 
-cd $DIR
+PATH="/etc:/bin:/usr/bin:/usr/etc:/usr/athena/etc"
+OUTFILE=/warehouse/transfers/moirausr/username_id.map.txt
 
-echo extracting $TAR_FILE
+# Alert if the output file doesn't exist
+test -r $OUTFILE || exit $MR_MISSINGFILE
 
-if [ ! -f $TAR_FILE ]; then
-	exit $MR_MISSINGFILE
-fi
+# Set the perms usefully
+chown root $OUTFILE
+chgrp 0 $OUTFILE
+chmod 644 $OUTFILE
 
-if zcat < $TAR_FILE | tar xvf -; then
-	echo Extract successful
-else
-	exit $MR_TARERR
-fi
-
-chmod 640 username_id.*
-
-rm -f $0
-echo removed self
+# cleanup
+test -f $0 && rm -f $0
 
 exit 0
